requirements.txt
# requirements.txt

faiss-cpu
sentence-transformers
PyMuPDF
numpy
pickle-mixin
